requirements.txt
matplotlib==3.1.3
opencv_python_headless==4.4.0.44
Flask==1.1.1
numpy==1.18.1
tensorflow-cpu==2.3.1
gunicorn==20.0.4
flask_cors==3.0.9